requirements.txt
Babel==2.8.0
bleach==3.1.0
colorama==0.4.3
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.8
django-import-export==2.0.2
django-markdown-filter==0.0.1
djangorestframework==3.11.0
et-xmlfile==1.0.1
jdcal==1.4.1
markdown2==2.3.8
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
six==1.13.0
sqlparse==0.3.0
tablib==1.1.0
webencodings==0.5.1
xlrd==1.2.0
xlwt==1.3.0
